Moved interruptible(...) implementation to platform-specific code
